Change data type to f64

+++ src/lib.rs
@@ -2,7 +2,7 @@ use std::{iter::zip, ops};
 
 #[derive(Debug)]
 pub struct Tensor {
-    pub data: Vec<i64>,
+    pub data: Vec<f64>,
     pub shape: Vec<usize>,
 }
 
@@ -12,7 +12,7 @@ impl ops::Add<Tensor> for Tensor {
     fn add(self, rhs: Tensor) -> Self::Output {
         assert_eq!(self.shape, rhs.shape);
 
-        let result: Vec<i64> = zip(self.data, rhs.data).map(|(x1, x2)| x1 + x2).collect();
+        let result: Vec<f64> = zip(self.data, rhs.data).map(|(x1, x2)| x1 + x2).collect();
 
         Tensor {
             data: result,
@@ -32,7 +32,7 @@ impl ops::Mul<Tensor> for Tensor {
 
         let rows: Vec<_> = self.data.chunks(self.shape[1]).collect();
 
-        let mut cols: Vec<Vec<i64>> = Vec::new();
+        let mut cols: Vec<Vec<f64>> = Vec::new();
 
         for (i, val) in rhs.data.iter().enumerate() {
             let remainder = i % rhs.shape[1];
@@ -59,14 +59,14 @@ impl ops::Mul<Tensor> for Tensor {
 }
 
 impl Tensor {
-    pub fn from_scalar(data: i64) -> Self {
+    pub fn from_scalar(data: f64) -> Self {
         Self {
             data: vec![data],
             shape: vec![],
         }
     }
 
-    pub fn new(data: Vec<i64>, shape: Vec<usize>) -> Self {
+    pub fn new(data: Vec<f64>, shape: Vec<usize>) -> Self {
         Self { data, shape }
     }
 }
@@ -77,29 +77,29 @@ mod tests {
 
     #[test]
     fn addition_1d() {
-        let a = Tensor::from_scalar(2);
-        let b = Tensor::from_scalar(3);
+        let a = Tensor::from_scalar(2.0);
+        let b = Tensor::from_scalar(3.0);
         let result = a + b;
 
-        assert_eq!(result.data, vec![5]);
+        assert_eq!(result.data, vec![5.0]);
     }
 
     #[test]
     fn addition_2d() {
-        let a = Tensor::new(vec![1, 2, 3, 4], vec![2, 2]);
-        let b = Tensor::new(vec![5, 6, 7, 8], vec![2, 2]);
+        let a = Tensor::new(vec![1.0, 2.0, 3.0, 4.0], vec![2, 2]);
+        let b = Tensor::new(vec![5.0, 6.0, 7.0, 8.0], vec![2, 2]);
         let result = a + b;
 
-        assert_eq!(result.data, vec![6, 8, 10, 12]);
+        assert_eq!(result.data, vec![6.0, 8.0, 10., 12.0]);
     }
 
     #[test]
     fn multiplication_2d() {
-        let a = Tensor::new(vec![1, 2, 3, 4, 5, 6], vec![2, 3]);
-        let b = Tensor::new(vec![1, 2, 3, 1, 2, 3], vec![3, 2]);
+        let a = Tensor::new(vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0], vec![2, 3]);
+        let b = Tensor::new(vec![1.0, 2.0, 3.0, 1.0, 2.0, 3.0], vec![3, 2]);
         let result = a * b;
 
-        assert_eq!(result.data, vec![13, 13, 31, 31]);
+        assert_eq!(result.data, vec![13.0, 13.0, 31.0, 31.0]);
         assert_eq!(result.shape, vec![2, 2]);
     }
 }
